+++ modeling/Token.java
@@ -1,7 +1,5 @@
 package modeling;
 
-import java.util.Objects;
-
 /**
  *
  * @author cristopher